Remove math operations from index
