+++ packages/wallet-specs/src/index.ts
@@ -1,7 +1,6 @@
 import { EventObject, SendAction, StateMachine, forwardTo } from 'xstate';
 import { Outcome, State } from '@statechannels/nitro-protocol';
 import { hashOutcome } from '@statechannels/nitro-protocol/lib/src/contract/outcome';
-import { bigNumberify } from 'ethers/utils';
 import { hashState } from '@statechannels/nitro-protocol/lib/src/contract/state';
 
 import { ChannelUpdated, IStore } from './store';
@@ -10,28 +9,6 @@ export { Store, IStore } from './store';
 export { SignedState } from './types';
 
 export { Channel, getChannelId } from '@statechannels/nitro-protocol';
-
-// This stuff should be replaced with some big number logic
-type numberish = string | number | undefined;
-type MathOp = (a: numberish, b: numberish) => string;
-export const add: MathOp = (a: numberish, b: numberish) =>
-  bigNumberify(a || 0)
-    .add(b || 0)
-    .toHexString();
-export const subtract: MathOp = (a: numberish, b: numberish) => {
-  const numA = bigNumberify(a || 0);
-  const numB = bigNumberify(b || 0);
-
-  if (numB.gt(numA)) {
-    throw new Error('Unsafe subtraction');
-  }
-  return numA.sub(numB).toHexString();
-};
-
-export const max: MathOp = (a: numberish, b: numberish) =>
-  Math.max(Number(a), Number(b)).toString();
-export const gt = (a: numberish, b: numberish) => Number(a) > Number(b);
-export const eq = (a: numberish, b: numberish) => Number(a) === Number(b);
 
 export const success: { type: 'final' } = { type: 'final' };
 export const failure: { type: 'final' } = { type: 'final' };
