+++ packages/wallet-specs/src/protocols/conclude-channel/protocol.ts
@@ -2,7 +2,8 @@ import { Machine } from 'xstate';
 
 import * as VirtualDefundingAsHub from '../virtual-defunding-as-hub/protocol';
 import * as VirtualDefundingAsLeaf from '../virtual-defunding-as-leaf/protocol';
-import { MachineFactory, FINAL, checkThat, add } from '../..';
+import { MachineFactory, FINAL, checkThat } from '../..';
+import { add } from '../../mathOps';
 import { IStore } from '../../store';
 import { getDetaAndInvoke } from '../../machine-utils';
 import { isIndirectFunding } from '../../ChannelStoreEntry';
